GUI works with comments and javaDoc

+++ src/Controller.java
@@ -39,24 +39,27 @@ public class Controller {
     @FXML
     CheckBox stem;
 
-    ArrayList<Pair<String,String>> dictionary;
-    PostingBuild pb;
+    ArrayList<Pair<String,String>> dictionary; //the loaded dictionary
+    PostingBuild pb;  //creates the dictionary and posting files
     boolean stemLoaded;
     boolean withoutStemLoaded;
 
     DirectoryChooser directoryChooser;
-    boolean firstTime;
 
-    public void initialize(){
+
+    public void initialize(){  //initializes all the variables
         this.directoryChooser = new DirectoryChooser();
         configuringDirectoryChooser(directoryChooser);
-        firstTime=true;
         dictionary=new ArrayList<>();
         stemLoaded=false;
         withoutStemLoaded=false;
         pb=new PostingFiles();
     }
 
+    /**
+     * @param event
+     * this function gets the corpus path from the text field using directory chooser
+     */
     public void getCorpusPath(ActionEvent event) {
         File dir = directoryChooser.showDialog(corpusButton.getScene().getWindow());
         if (dir != null) {
@@ -66,7 +69,11 @@ public class Controller {
         }
     }
 
-    public void getDictionaryPath(ActionEvent event) {
+    /**
+     * @param event
+     * This function gets the dictionary path from the text field using directory chooser
+     */
+    public void getDictionaryPath(ActionEvent event) {  //gets the dictionary path using directory chooser
         File dir = directoryChooser.showDialog(dictionaryButton.getScene().getWindow());
         if (dir != null) {
             dictionaryPath.setText(dir.getAbsolutePath());
@@ -75,30 +82,38 @@ public class Controller {
         }
     }
 
+    /**
+     * @param event
+     * This function creates the dictionary after you press the Create dictionary button
+     */
     public void createDictionary(ActionEvent event) {
         String path=dictionaryPath.getText();
         String dictPath=path+"/dictionary";
-        if(corpusPath.getText().equals("")&&path.equals("")) {
+        if(corpusPath.getText().equals("")&&path.equals("")) {   //checks if all the paths exists
             openError("You didnt enter the corpus path or the wanted Posting files path.");
             return;
         }
-        new File(dictPath).mkdir();
-        pb.startBuildingStock(Paths.get(dictPath),Paths.get(corpusPath.getText()));
+        new File(dictPath).mkdir(); //creates the dictionary directory
+        pb.startBuildingStock(Paths.get(dictPath),Paths.get(corpusPath.getText()));  //creates the dictionary and the posting files
         openInformation("Dictionary created successfully!");
     }
 
+    /**
+     * @param event
+     * This function loads the dictionary to the main memory after clicking the Load dictionary button
+     */
     public void loadDictionary(ActionEvent event){
         File dictionary;
-        this.dictionary.clear();
+        this.dictionary.clear();  //deletes the former dictionary if loaded
         if(stem.isSelected())
             dictionary = new File(dictionaryPath.getText()+"/dictionary/withStem AllDictionary.txt");
         else
             dictionary = new File(dictionaryPath.getText()+"/dictionary/withoutStem AllDictionary.txt");
-        if(dictionaryPath.getText().equals("")||!dictionary.exists()) {
+        if(dictionaryPath.getText().equals("")||!dictionary.exists()) {  //checks if the dictionary exists
             openError("You didnt enter one of the paths or didnt create the dictionary.");
             return;
         }
-        try
+        try  //reads the dictionary
         {
             BufferedReader reader;
             if(stem.isSelected()) {
@@ -128,18 +143,20 @@ public class Controller {
     }
 
 
-
-
+    /**
+     * @param event
+     * This function opens the dictionary in a new window after clicking the Open dictionary button
+     */
     public void openDictionary(ActionEvent event) {
-        if(!stemLoaded && !withoutStemLoaded) {
+        if(!stemLoaded && !withoutStemLoaded) {   //checks if a dictionary was loaded
             openError("You didnt load the dictionary.");
             return;
         }
-        if(stem.isSelected()&&!stemLoaded){
+        if(stem.isSelected()&&!stemLoaded){  //checks if the right dictionary was loaded
             openError("You didnt load the dictionary.");
             return;
         }
-        else if(!stem.isSelected()&&!withoutStemLoaded){
+        else if(!stem.isSelected()&&!withoutStemLoaded){ //checks if the right dictionary was loaded
             openError("You didnt load the dictionary.");
             return;
         }
@@ -161,12 +178,17 @@ public class Controller {
         }
 
 
+    /**
+     * @param event
+     * This function resets the dictionary and deletes all the files that were created with him
+     * and also cleans the main memory
+     */
     public void resetDictionary(ActionEvent event) {
         String path = dictionaryPath.getText();
         String dictPath = dictionaryPath.getText()+"/dictionary";
         File dictionary = new File(dictPath+"/withoutStem AllDictionary.txt");
-        if(!path.equals("")&&dictionary.exists()) {
-            pb.deleteFolderInformation(Paths.get(dictPath));
+        if(!path.equals("")&&dictionary.exists()) {   //checks if the dictionary exists and the path was entered
+            pb.deleteFolderInformation(Paths.get(dictPath));  //deletes all the files that were created
             new File(dictPath).delete();
             dictionaryPath.clear();
             corpusPath.clear();
@@ -178,12 +200,21 @@ public class Controller {
     }
 
 
+    /**
+     * @param info A string that contains the alert information
+     * This function opens an information alert with a message according the the call
+     */
     public void openInformation(String info){
             Alert missing=new Alert(Alert.AlertType.INFORMATION);
             missing.setContentText(info);
             missing.showAndWait();
         }
 
+
+    /**
+     * @param error A string that contains the alert information
+     * This function opens an error alert with a message according the the call
+     */
     public void openError(String error){
         Alert missing=new Alert(Alert.AlertType.ERROR);
         missing.setContentText(error);
@@ -191,7 +222,10 @@ public class Controller {
     }
 
 
-
+    /**
+     * @param directoryChooser
+     * This function configures the directory chooser used to get the paths
+     */
     private void configuringDirectoryChooser(DirectoryChooser directoryChooser) {
         // Set title for DirectoryChooser
         directoryChooser.setTitle("Select Some Directories");
